Move parse-event back into the events file as it's not used anywhere else.
